feat: optimized register forms, refactored components structure
