+++ src/app/(non-authenticated-routes)/register/styles.module.scss
@@ -45,14 +45,14 @@
     cursor: pointer;
     display: grid;
     font-family: $m;
-    font-size: 24px;
+    font-size: 22px;
     font-weight: 800;
-    height: 55px;
+    height: 45px;
     padding: 13px;
     place-content: center;
     text-align: center;
     transition: 0.2s;
-    width: 238px;
+    width: 218px;
 }
 
 #candidate  {
