+++ src/presentation/components/button/styles.module.scss
@@ -63,3 +63,7 @@
         height: 35px;
     }
 }
+
+.large  {
+    font-size: 20px;
+}
