+++ src/presentation/components/input/styles.module.scss
@@ -8,7 +8,7 @@
 .label  {
     color: $light-blue;
     font-family: $m;
-    font-size: 24px;
+    font-size: 20px;
     font-weight: 600;
 }
 
@@ -21,12 +21,12 @@
     border-radius: 10px;
     border: none;
     color: $white-gray;
-    font-family: $m;
+    font-family: sans-serif;
     font-size: 18px;
-    height: 60px;
+    height: 50px;
     padding: 10px;
     outline: none;
-    width: 655px;
+    width: 305px;
 
     @media  (max-width: 1440px) {
         height: 50px;
