+++ src/presentation/components/register/company-register/styles.module.scss
@@ -1,12 +1,15 @@
 @import '@/app/globals.scss';
 
 .register_company_form  {
+    align-items: center;
     display: flex;
     flex-direction: column;
+    gap: 30px;
 }
 
 .form_rows  {
     display: flex;
     flex-direction: row;
     gap: 72px;
+    width: 100%;
 }
