+++ src/presentation/components/select/styles.module.scss
@@ -8,8 +8,12 @@
 .label  {
     color: $light-blue;
     font-family: $m;
-    font-size: 24px;
+    font-size: 20px;
     font-weight: 600;
+}
+
+.uppercase  {
+    text-transform: uppercase;
 }
 
 .select  {
@@ -17,12 +21,11 @@
     border-radius: 10px;
     border: none;
     color: $white-gray;
-    font-family: $m;
+    font-family: sans-serif;
     font-size: 18px;
-    height: 60px;
+    height: 50px;
     padding: 10px;
     outline: none;
-    width: 655px;
 
     @media  (max-width: 1440px) {
         height: 50px;
